+++ apps/client/src/app/components/rule/rule-settings-dialog/rule-settings-dialog.component.ts
@@ -3,7 +3,7 @@ import { GfValueComponent } from '@ghostfolio/ui/value';
 
 import { CommonModule } from '@angular/common';
 import { Component, Inject } from '@angular/core';
-import { FormsModule } from '@angular/forms';
+import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import {
   MAT_DIALOG_DATA,
@@ -17,7 +17,7 @@ import { RuleSettingsDialogParams } from './interfaces/interfaces';
 @Component({
   imports: [
     CommonModule,
-    FormsModule,
+    ReactiveFormsModule,
     GfValueComponent,
     MatButtonModule,
     MatDialogModule,
@@ -29,9 +29,23 @@ import { RuleSettingsDialogParams } from './interfaces/interfaces';
 })
 export class GfRuleSettingsDialogComponent {
   public settings: XRayRulesSettings['AccountClusterRiskCurrentInvestment'];
+  public form: FormGroup;
 
   public constructor(
     @Inject(MAT_DIALOG_DATA) public data: RuleSettingsDialogParams,
-    public dialogRef: MatDialogRef<GfRuleSettingsDialogComponent>
-  ) {}
+    public dialogRef: MatDialogRef<GfRuleSettingsDialogComponent>,
+    private formBuilder: FormBuilder
+  ) {
+    this.form = this.formBuilder.group({
+      thresholdMax: [this.data.settings?.thresholdMax],
+      thresholdMin: [this.data.settings?.thresholdMin]
+    });
+  }
+
+  public onSave() {
+    this.dialogRef.close({
+      ...this.data.settings,
+      ...this.form.value
+    });
+  }
 }